+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -3,6 +3,19 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v0.40.0
+-------------------------------
+*January 20, 2021*
+
+### Added
+- `validations` mixin.
+- `Guest` component validations.
+- `VALIDATIONS` constant.
+
+### Changed
+- `updateAddressDetails` and `updateCustomerDetails` store actions.
+
+
 v0.39.0
 -------------------------------
 *January 18, 2021*
@@ -17,6 +30,7 @@ v0.38.0
 
 ### Changed
 - Use the latest version of `f-form-field`.
+
 
 v0.37.0
 -------------------------------